refactor: 토스트 위치 이동
