+++ components/common/Toast/ToastWithSignupLink.stories.tsx
@@ -1,7 +1,7 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react';
 
 import SignUpLink from '@/components/projects/upload/SignUpLink';
-import { ToastProvider } from '@/components/projects/upload/ToastProvider';
+import { ToastProvider } from '@/components/common/Toast';
 
 const ToastWithSignUpLink = () => {
   return (
